Features of Medicinal Plants

feature description of medicinal plant

Cvariable_name

variable description

ID

Id for the plant(Primary key)

Sinhala_Name

Sinhala name of the plant(Local name)

Family_Name

Family name of the plant

Scientific_Name

Scientific name of the plant

Shape

Shape of the leaf:1 = Heart, 2 = Round, 3 = Needle, 4 = Spear shaped/Diamond, 5 = Hand shaped (In Main category Diamond), 6 = Wedge shaped (In Main category Diamond), 7 = Scale-like shaped, 8 = Simple round

Arrangements

Arrangements of the leaves on the main stem:, 1 = Simple, 2 = Compound

Bipinnately_compound

Is the compound type leaf bipinnately arranged or not?: 0 = False, 1 = True

Pinnately_compound

Is the compound type leaf pinnately arranged or not?: 0 = False, 1 = True

Palmately_compound

Is the compound type leaf palmately arranged or not?: 0 = False, 1 = True

Edges

Edge type of the leaf: 1 = Smooth, 2 = Toothed, 3 = Lobed, 4 = Crenate

Uniform_background

Is the background of the leaf uniform or not?: 0 = False, 1 = True

Red_Margin

Is there a red margin of the leaf or not?: 0 = False, 1 = True

Shaded_margin

Is there a shaded margin of the leaf or not?: 0 = False, 1 = True

White_Shading

Is there a white shading on the leaf or not?: 0 = False, 1 = True

Red_Shading

Is there a red shading on the leaf or not?: 0 = False, 1 = True

White_line

Is there a white line on the leaf or not?: 0 = False, 1 = True

Green_leaf

Is the leaf green colored or not?: 0 = False, 1 = True

Red_leaf

Is the leaf red colored or not?: 0 = False, 1 = True

Veins

Vein category of the leaf: 1 = Parallel, 2 = Pinnate, 3 = Palmate

Arrangement_on_the_stem

Category of leaf arrangements in the plant: 1 = Alternate, 2 = Opposite, 3 = Whorled

Leaf_Apices

Apice type of the leaf: 1 = Acute, 2 = Obtuse, 3 = Rounded, 4 = Cuspidate, 5 = Mucronulate, 6 = Truncate, 7 = Truncate

Leaf_Base

Base type of the leaf: 1 = Acuate, 2 = Rounded, 3 = Auriculate, 4 = Hastate, 5 = Gradually tapering, 6 = Truncate, 7 = Cordate, 8 = Attenuate, 9 = Tumcate, 10 = Cuneate, 11 = Obtuse, 12 = Cleft

medlea

Format

An object of class data.frame with 471 rows and 22 columns.